pyproject.toml
[project]
name = "prime_vfv"
version = "0.0.0"
description = "PRIME + Vul-Fix-Vul research"
authors = [{ name = "Nicholas M. Synovic", email = "nicholas.synovic@gmail.com" }]
requires-python = "~=3.13"
readme = "README.md"
keywords = ["Template"]
classifiers = [
    "Development Status :: 1 - Planning",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3.13"
]
dependencies = [
    "click>=8.3.0",
    "gitpython>=3.1.45",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "progress>=1.6.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.43",
]

[project.urls]
Homepage = "https://github.com/NicholasSynovic/template_python"
Repository = "https://github.com/NicholasSynovic/template_python"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
